requirements.txt
altgraph==0.17
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
importlib-metadata==3.10.0
macholib==1.14
pycparser==2.20
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pyOpenSSL==20.0.1
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
soupsieve==2.2.1
typing-extensions==3.7.4.3
Unidecode==1.2.0
urllib3==1.26.4
zipp==3.4.1
